<==== Mercurial data ====>
Branch: feature_3334_time_sync

+++ client/src/ui/workbench/handlers/workbench_screenshot_handler.cpp
@@ -23,11 +23,15 @@
 #include <ui/graphics/items/resource/media_resource_widget.h>
 #include <ui/dialogs/custom_file_dialog.h>
 #include <ui/dialogs/progress_dialog.h>
+#include <ui/dialogs/workbench_state_dependent_dialog.h>
+
+#include <ui/workbench/workbench_context.h>
 #include <ui/workbench/workbench_item.h>
 
 #include <utils/common/string.h>
 #include <utils/common/environment.h>
 #include <utils/common/warnings.h>
+#include "transcoding/filters/fisheye_image_filter.h"
 
 //#define QN_SCREENSHOT_DEBUG
 #ifdef QN_SCREENSHOT_DEBUG
@@ -254,11 +258,14 @@ void QnWorkbenchScreenshotHandler::at_takeScreenshotAction_triggered() {
             allowedFormatFilter += jpegFileFilter;
         }
 
-        QScopedPointer<QnCustomFileDialog> dialog(new QnCustomFileDialog(
-            mainWindow(),
-            tr("Save Screenshot As..."),
-            suggestion,
-            allowedFormatFilter
+        bool wasLoggedIn = !context()->user().isNull();
+
+        QScopedPointer<QnCustomFileDialog> dialog(
+            new QnWorkbenchStateDependentDialog<QnCustomFileDialog> (
+                mainWindow(),
+                tr("Save Screenshot As..."),
+                suggestion,
+                allowedFormatFilter
         ));
 
         dialog->setFileMode(QFileDialog::AnyFile);
@@ -275,6 +282,11 @@ void QnWorkbenchScreenshotHandler::at_takeScreenshotAction_triggered() {
         dialog->addWidget(tr("Timestamp:"), comboBox);
 
         if (!dialog->exec())
+            return;
+
+        /* Check if we were disconnected (server shut down) while the dialog was open. 
+         * Skip this check if we were not logged in before. */
+        if (wasLoggedIn && !context()->user())
             return;
 
         QString fileName = dialog->selectedFile();
@@ -298,6 +310,11 @@ void QnWorkbenchScreenshotHandler::at_takeScreenshotAction_triggered() {
             }
         }
 
+        /* Check if we were disconnected (server shut down) while the dialog was open. 
+         * Skip this check if we were not logged in before. */
+        if (wasLoggedIn && !context()->user())
+            return;
+
         if (QFile::exists(fileName) && !QFile::remove(fileName)) {
             QMessageBox::critical(
                 mainWindow(),
@@ -307,6 +324,11 @@ void QnWorkbenchScreenshotHandler::at_takeScreenshotAction_triggered() {
             );
             return;
         }
+
+        /* Check if we were disconnected (server shut down) while the dialog was open. 
+         * Skip this check if we were not logged in before. */
+        if (wasLoggedIn && !context()->user())
+            return; //TODO: #GDM triple check...
 
         parameters.filename = fileName;
         parameters.timestampPosition = static_cast<Qn::Corner>(comboBox->itemData(comboBox->currentIndex()).value<int>());
@@ -365,8 +387,7 @@ void QnWorkbenchScreenshotHandler::at_imageLoaded(const QImage &image) {
     
     int panoFactor = (parameters.mediaDewarpingParams.enabled && parameters.itemDewarpingParams.enabled) ? parameters.itemDewarpingParams.panoFactor : 1;
     if (panoFactor > 1)
-        resized = resized.scaled(resized.width() * panoFactor, resized.height());
-    
+        resized = resized.scaled(QnFisheyeImageFilter::getOptimalSize(resized.size(), parameters.itemDewarpingParams));
 
     QScopedPointer<CLVideoDecoderOutput> frame(new CLVideoDecoderOutput(resized));
     if (parameters.imageCorrectionParams.enabled) {
@@ -400,7 +421,7 @@ void QnWorkbenchScreenshotHandler::at_imageLoaded(const QImage &image) {
 
 void QnWorkbenchScreenshotHandler::showProgressDelayed(const QString &message) {
     if (!m_screenshotProgressDialog) {
-        m_screenshotProgressDialog = new QnProgressDialog(mainWindow());
+        m_screenshotProgressDialog = new QnWorkbenchStateDependentDialog<QnProgressDialog>(mainWindow());
         m_screenshotProgressDialog->setWindowTitle(tr("Saving...")); // TODO: #string_freeze replace with "Saving Screenshot"
         m_screenshotProgressDialog->setInfiniteProgress();
         // TODO: #dklychkov ensure concurrent screenshot saving is ok and disable modality
